Refactor location hierarchy forms to use abstract base class
